+++ internal/config/config_scenarios_test.go
@@ -21,6 +21,7 @@ var configScenarios = []dsl.Scenario{
 		assert.Equal(t, "Logging Enabled default", cfg.Logging.Enabled, false)
 		assert.Equal(t, "Logging Level default", cfg.Logging.Level, "INFO")
 		assert.Equal(t, "Z80 instructions logging default", cfg.Logging.Z80.Instructions, true)
+		assert.Equal(t, "Z80 tape logging default", cfg.Logging.Z80.Tape, true)
 		assert.Equal(t, "Border Color default", cfg.Display.Border.Color, "#D6EFC9")
 		assert.Equal(t, "Border Width default", cfg.Display.Border.Width, 15)
 		assert.Equal(t, "Toolbar Color default", cfg.Display.Toolbar.Color, "#D6CDC9")
@@ -41,6 +42,7 @@ var configScenarios = []dsl.Scenario{
 				Level:   "DEBUG",
 				Z80: Z80LoggingConfig{
 					Instructions: false,
+					Tape:         false,
 				},
 			},
 			Display: DisplayConfig{
@@ -66,6 +68,7 @@ var configScenarios = []dsl.Scenario{
 		assert.Equal(t, "Logging Enabled", cfg.Logging.Enabled, true)
 		assert.Equal(t, "Logging Level", cfg.Logging.Level, "DEBUG")
 		assert.Equal(t, "Z80 instructions logging", cfg.Logging.Z80.Instructions, false)
+		assert.Equal(t, "Z80 tape logging", cfg.Logging.Z80.Tape, false)
 		assert.Equal(t, "Border Color", cfg.Display.Border.Color, "#FF0000")
 		assert.Equal(t, "Border Width", cfg.Display.Border.Width, 30)
 		assert.Equal(t, "Toolbar Color", cfg.Display.Toolbar.Color, "#00FF00")
@@ -89,5 +92,6 @@ var configScenarios = []dsl.Scenario{
 		cfg, err := Load(configPath)
 		assert.Equal(t, "Load error should be nil", err, nil)
 		assert.Equal(t, "Z80 instructions logging default", cfg.Logging.Z80.Instructions, true)
+		assert.Equal(t, "Z80 tape logging default", cfg.Logging.Z80.Tape, true)
 	}),
 }